-save transactions in db untest

+++ src/main/java/com/happyzombie/springinitializr/api/TransactionBaseInfo.java
@@ -2,7 +2,6 @@ package com.happyzombie.springinitializr.api;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
@@ -11,7 +10,7 @@ public class TransactionBaseInfo {
     private String signerId;
     private String receiverId;
     private String blockHash;
-    private BigInteger blockTimestamp;
+    private Long blockTimestamp;
     private Integer transactionIndex;
     private List<Action> actions;
 }
